push_swap

+++ libft/ft_atoi.c
@@ -10,10 +10,10 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	j;
+	int		i;
+	long	j;
 
 	i = 1;
 	j = 0;
